+++ ethereum/blockwatch/block_watcher_test.go
@@ -48,14 +48,14 @@ func TestWatcher(t *testing.T) {
 	for i := 0; i < fakeClient.NumberOfTimesteps(); i++ {
 		scenarioLabel := fakeClient.GetScenarioLabel()
 
-		err = watcher.syncChain()
+		err = watcher.SyncToLatestBlock()
 		if strings.HasPrefix(scenarioLabel, "ERROR") {
 			require.Error(t, err)
 		} else {
 			require.NoError(t, err)
 		}
 
-		retainedBlocks, err := watcher.GetAllRetainedBlocks()
+		retainedBlocks, err := watcher.getAllRetainedBlocks()
 		require.NoError(t, err)
 		expectedRetainedBlocks := fakeClient.ExpectedRetainedBlocks()
 		assert.Equal(t, expectedRetainedBlocks, retainedBlocks, scenarioLabel)
@@ -195,7 +195,7 @@ func TestGetSubBlockRanges(t *testing.T) {
 	}
 }
 
-func TestSyncToLatestBlockLessThan128Missed(t *testing.T) {
+func TestFastSyncToLatestBlockLessThan128Missed(t *testing.T) {
 	// Fixture will return block 132 as the tip of the chain (127 blocks from block 5)
 	fakeClient, err := newFakeClient("testdata/fake_client_fast_sync_fixture.json")
 	require.NoError(t, err)
@@ -219,7 +219,7 @@ func TestSyncToLatestBlockLessThan128Missed(t *testing.T) {
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	blocksElapsed, err := watcher.SyncToLatestBlock(ctx)
+	blocksElapsed, err := watcher.FastSyncToLatestBlock(ctx)
 	require.NoError(t, err)
 	assert.Equal(t, 127, blocksElapsed)
 
@@ -230,7 +230,7 @@ func TestSyncToLatestBlockLessThan128Missed(t *testing.T) {
 	assert.Equal(t, big.NewInt(132), headers[0].Number)
 }
 
-func TestSyncToLatestBlockMoreThanOrExactly128Missed(t *testing.T) {
+func TestFastSyncToLatestBlockMoreThanOrExactly128Missed(t *testing.T) {
 	// Fixture will return block 133 as the tip of the chain (128 blocks from block 5)
 	fakeClient, err := newFakeClient("testdata/fake_client_reset_fixture.json")
 	require.NoError(t, err)
@@ -254,7 +254,7 @@ func TestSyncToLatestBlockMoreThanOrExactly128Missed(t *testing.T) {
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	blocksElapsed, err := watcher.SyncToLatestBlock(ctx)
+	blocksElapsed, err := watcher.FastSyncToLatestBlock(ctx)
 	require.NoError(t, err)
 	assert.Equal(t, 128, blocksElapsed)
 
@@ -264,7 +264,7 @@ func TestSyncToLatestBlockMoreThanOrExactly128Missed(t *testing.T) {
 	require.Len(t, headers, 0)
 }
 
-func TestSyncToLatestBlockNoneMissed(t *testing.T) {
+func TestFastSyncToLatestBlockNoneMissed(t *testing.T) {
 	// Fixture will return block 5 as the tip of the chain
 	fakeClient, err := newFakeClient("testdata/fake_client_basic_fixture.json")
 	require.NoError(t, err)
@@ -288,7 +288,7 @@ func TestSyncToLatestBlockNoneMissed(t *testing.T) {
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
-	blocksElapsed, err := watcher.SyncToLatestBlock(ctx)
+	blocksElapsed, err := watcher.FastSyncToLatestBlock(ctx)
 	require.NoError(t, err)
 	assert.Equal(t, blocksElapsed, 0)