+++ ethereum/ratelimit/rate_limiter_test.go
@@ -24,8 +24,8 @@ var (
 	minExpectedDelay                      = time.Duration(1000) / time.Duration(maxRequestsPerSecond) * time.Millisecond
 	defaultCheckpointInterval             = 1 * time.Minute
 	expectedMaxElapsedTimeForFirstRequest = 1 * time.Millisecond
-	expectedDeltaMinExpectedDelay         = 15 * time.Millisecond
-	expectedDeltaMaxExpectedDelay         = 15 * time.Millisecond
+	expectedDeltaMinExpectedDelay         = 20 * time.Millisecond
+	expectedDeltaMaxExpectedDelay         = 20 * time.Millisecond
 )
 
 // Scenario1: Mesh starts X seconds after UTC midnight (start of next UTC day) and